Switch data type of timestamp from long to long long

+++ src/datapoint.cpp
@@ -4,7 +4,7 @@ DataPoint::DataPoint(){
   this->initialized = false;
 }
 
-DataPoint:: DataPoint(const long timestamp, const DataPointType data_type, const VectorXd raw){
+DataPoint:: DataPoint(const long long timestamp, const DataPointType data_type, const VectorXd raw){
   this->set(timestamp, data_type, raw);
 }
